requirements_dev.txt
google-cloud-pubsub
google-cloud-firestore
apache-beam[gcp]